mais refatorações
