+++ semana19/LaBook/src/model/User.ts
@@ -1,0 +1,6 @@
+export type User = {
+    id: string,
+    name: string,
+    email: string,
+    password: string
+}
